feat: added betting logic (Method 2), ledger endpoints, and minimum bet validation

+++ backend/app/models.py
@@ -22,7 +22,7 @@ class User(Base):
     id = Column(Integer, primary_key=True, index=True)
     email = Column(String, unique=True, index=True, nullable=False)
     hashed_password = Column(String, nullable=False)
-    balance = Column(Float, default=0.0)
+    balance = Column(Integer, default=0)
     created_at = Column(DateTime, default=datetime.utcnow)
     is_verified = Column(Boolean, default=False)
     is_admin = Column(Boolean, default=False)
